refactor dynamic handlers
